Change DecimalField to IntegerField on LecturaSignos
Fixes #217

+++ hospinet/clinique/models.py
@@ -118,11 +118,9 @@ class LecturaSignos(TimeStampedModel):
     pulso = models.IntegerField()
     temperatura = models.DecimalField(decimal_places=2, max_digits=8,
                                       null=True)
-    presion_sistolica = models.DecimalField(decimal_places=2, max_digits=8,
-                                            null=True)
-    presion_diastolica = models.DecimalField(decimal_places=2, max_digits=8,
-                                             null=True)
-    respiracion = models.DecimalField(decimal_places=2, max_digits=8, null=True)
+    presion_sistolica = models.IntegerField(null=True)
+    presion_diastolica = models.IntegerField(null=True)
+    respiracion = models.IntegerField(null=True)
     presion_arterial_media = models.CharField(max_length=200, blank=True)
 
     def get_absolute_url(self):
